# be/docs/swagger.yaml
basePath: /api
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.ErrorData'
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  models.ErrorData:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@form.com
    name: API Support
  description: API for form application with JWT authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: form API
  version: "1.0"
paths:
...
  /forms/{form_id}/analytics:
    get:
      consumes:
      - application/json
      description: Retrieve detailed analytics including summary statistics, score
        distribution, time analytics, and question performance
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Start date for analytics period (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: End date for analytics period (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - description: Group responses by time period
        enum:
        - day
        - week
        - month
        in: query
        name: group_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get comprehensive analytics for a form
      tags:
      - analytics
  /forms/{form_id}/analytics/questions:
    get:
      consumes:
      - application/json
      description: Retrieve detailed analytics for each question including answer
        distribution and performance metrics
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get question-level analytics for a form
      tags:
      - analytics
  /forms/{form_id}/analytics/sections:
    get:
      consumes:
      - application/json
      description: Retrieve analytics for each section including average scores and
        completion rates
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get section-level analytics for a form
      tags:
      - analytics
  /forms/{form_id}/analytics/trends:
    get:
      consumes:
      - application/json
      description: Retrieve trend data showing responses over time with grouping options
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Group responses by time period
        enum:
        - day
        - week
        - month
        in: query
        name: group_by
        required: true
        type: string
      - description: Start date for trend analysis (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: End date for trend analysis (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get response trends over time for a form
      tags:
      - analytics
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"